Implement review comments
