+++ cebra/solver/base.py
@@ -405,33 +405,11 @@ class Solver(abc.ABC, cebra.io.HasDevice):
         Yields:
             The parameters of the model.
         """
-        raise NotImplementedError
+        for parameter in self.model.parameters():
+            yield parameter
 
-    def _compute_features(
-        self,
-        batch: cebra.data.Batch,
-        model: Optional[torch.nn.Module] = None
-    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """Compute the features of the reference, positive and negative samples.
-
-        Args:
-            batch: The input data, not necessarily aligned across the batch
-                dimension. This means that ``batch.index`` specifies the map
-                between reference/positive samples, if not equal ``None``.
-            model: The model to use for inference.
-                If not provided, the model of the solver is used.
-
-        Returns:
-            Tuple of reference, positive and negative features.
-        """
-        if model is None:
-            model = self.model
-
-        batch.to(self.device)
-        ref = model(batch.reference)
-        pos = model(batch.positive)
-        neg = model(batch.negative)
-        return ref, pos, neg
+        for parameter in self.criterion.parameters():
+            yield parameter
 
     def _get_loader(self, loader):
         return ProgressBar(
